moving warning to crt
