+++ lib/native/auth.ts
@@ -16,7 +16,7 @@ import crt_native from './binding';
 import { CrtError } from './error';
 import { HttpRequest, HttpProxyOptions } from './http';
 import {ClientBootstrap, ClientTlsContext} from './io';
-import node_deprecation_warning from '../common/node_deprecation_warning';
+import {emitNodeDeprecationWarning } from './crt'
 
 export {AwsSigningConfigBase} from "../common/auth";
 
@@ -141,7 +141,7 @@ export class AwsCredentialsProvider extends crt_native.AwsCredentialsProvider {
      * @returns a new credentials provider using default credentials resolution rules
      */
     static newDefault(bootstrap: ClientBootstrap | undefined = undefined): AwsCredentialsProvider {
-        node_deprecation_warning.emitWarning();
+        emitNodeDeprecationWarning()
         return super.newDefault(bootstrap != null ? bootstrap.native_handle() : null);
     }
 
@@ -155,7 +155,7 @@ export class AwsCredentialsProvider extends crt_native.AwsCredentialsProvider {
      * @returns a new credentials provider that will return a fixed set of AWS credentials
      */
     static newStatic(access_key: crt_native.StringLike, secret_key: crt_native.StringLike, session_token?: crt_native.StringLike): AwsCredentialsProvider {
-        node_deprecation_warning.emitWarning();
+        emitNodeDeprecationWarning()
         return super.newStatic(access_key, secret_key, session_token);
     }
 
@@ -168,7 +168,7 @@ export class AwsCredentialsProvider extends crt_native.AwsCredentialsProvider {
      * @returns a new credentials provider that returns credentials sourced from the AWS Cognito Identity service
      */
     static newCognito(config: CognitoCredentialsProviderConfig): AwsCredentialsProvider {
-        node_deprecation_warning.emitWarning();
+        emitNodeDeprecationWarning()
         if (config == null || config == undefined) {
             throw new CrtError("AwsCredentialsProvider newCognito: Cognito config not defined");
         }
@@ -187,7 +187,7 @@ export class AwsCredentialsProvider extends crt_native.AwsCredentialsProvider {
      * @returns a new credentials provider that returns credentials sourced from the AWS X509 service
      */
     static newX509(config : X509CredentialsConfig): AwsCredentialsProvider {
-        node_deprecation_warning.emitWarning();
+        emitNodeDeprecationWarning()
         if (config == null || config == undefined) {
             throw new CrtError("AwsCredentialsProvider newX509: X509 config not defined")
         }
